requirements.txt
fastapi
uvicorn
firebase-admin
pinecone-client
transformers
torch
PyPDF2
requests
python-dotenv
pydantic
